+++ src/const.py
@@ -1,3 +1,6 @@
-APP_ROOT_DIR = "/app/"
-DATA_FILE_EXT = ".json"
-UTC = "UTC"
+from pathlib import Path
+
+
+APP_ROOT_DEFAULT: Path = Path("/app/")
+DATA_FILE_EXT: str = ".json"
+UTC: str = "UTC"