+++ src/tasks/lunartime.py
@@ -214,7 +214,7 @@ class LunarTimeTask(object):
 
 
     def _loadLunarTime(self) -> Dict:
-        filePath = Path(self._data_dir + self._TASK_NAME + DATA_FILE_EXT)
+        filePath = Path.joinpath(self._data_dir, self._TASK_NAME + DATA_FILE_EXT)
         filePath.touch(exist_ok=True)
         with open(filePath, 'r') as fp:
             try: 
@@ -226,7 +226,7 @@ class LunarTimeTask(object):
 
 
     def _saveLunarTime(self, lunar_time: Dict) -> None:
-        fw = open(self._data_dir + self._TASK_NAME + DATA_FILE_EXT, 'w+')
+        fw = open(Path.joinpath(self._data_dir, self._TASK_NAME + DATA_FILE_EXT), 'w+')
         json.dump(lunar_time, fw, default=self._dumpConverter, indent=2)
         fw.close()