+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## Unreleased
 ### Added
 - The ability to deploy the application via an image in a docker container
+- The ability to specify the path to the application's root directory
 
 ## Release 0.2.1 [2022-02-01]
 ### Fixed